pyproject.toml
[project]
name = "projeto-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.135.2",
    "google-genai>=1.73.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pwdlib[bcrypt]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
